feat: added checks for Interfaces
