--- md/syntaxrules-explanation-identifier_matches_filename.md
@@ -1,3 +1,0 @@
-Module Identifier should have the same name as the file it's in.
-
-```module Bar``` should be in ```some/path/to/Bar.sv```
