+++ md/syntaxrules-explanation-interface_identifier_matches_filename.md
@@ -1,0 +1,3 @@
+Interface Identifier should have the same name as the file it's in.
+
+```interface Bar``` should be in ```some/path/to/Bar.sv```
