--- testcases/syntaxrules/fail/identifier_matches_filename.sv
@@ -1,2 +1,0 @@
-module Bar;
-endmodule
