+++ testcases/syntaxrules/fail/interface_identifier_matches_filename.sv
@@ -1,0 +1,2 @@
+interface Bar;
+endinterface
