--- testcases/syntaxrules/pass/identifier_matches_filename.sv
@@ -1,2 +1,0 @@
-module syntaxrules_identifier_matches_filename_pass_1of1;
-endmodule
