+++ testcases/syntaxrules/pass/interface_identifier_matches_filename.sv
@@ -1,0 +1,2 @@
+interface syntaxrules_interface_identifier_matches_filename_pass_1of1;
+endinterface
